requirements.txt
customtkinter
ffmpeg-python
opencv-python
pillow
imageio-ffmpeg
pyinstaller
